# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: TourVision API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /user/profile:
    get:
      operationId: getUserProfile
      tags: [user]
      summary: Get current user profile
      responses:
        "200":
          description: User profile
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserProfile"
    put:
      operationId: updateUserProfile
      tags: [user]
      summary: Update user profile
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateUserProfileBody"
      responses:
        "200":
          description: Updated profile
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserProfile"
...
  /user/onboarding:
    post:
      operationId: completeOnboarding
      tags: [user]
      summary: Complete user onboarding
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/OnboardingBody"
      responses:
        "200":
          description: Onboarding completed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
...
  /subscribe:
    post:
      operationId: subscribeNewsletter
      tags: [user]
      summary: Subscribe to newsletter (public)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/NewsletterSubscribeBody"
      responses:
        "200":
          description: Subscribed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
components:
  schemas:
...
    SuccessResponse:
      type: object
      properties:
        success:
          type: boolean
        message:
          type: string
      required:
        - success
...
    UserProfile:
      type: object
      properties:
        id:
          type: string
        fullName:
          type: string
        email:
          type: string
        avatarUrl:
          type: string
          nullable: true
        accountType:
          type: string
          enum: [agent, host, developer, manager, other]
        country:
          type: string
          nullable: true
        whatsappNumber:
          type: string
          nullable: true
        subscriptionTier:
          type: string
          enum: [free, pro, unlimited]
        subscriptionStatus:
          type: string
          nullable: true
        toursThisMonth:
          type: integer
        totalTours:
          type: integer
        onboardingCompleted:
          type: boolean
        createdAt:
          type: string
      required:
        - id
        - fullName
        - email
        - accountType
        - subscriptionTier
        - toursThisMonth
        - totalTours
        - onboardingCompleted
        - createdAt
    UpdateUserProfileBody:
      type: object
      properties:
        fullName:
          type: string
        accountType:
          type: string
          enum: [agent, host, developer, manager, other]
        country:
          type: string
        whatsappNumber:
          type: string
    OnboardingBody:
      type: object
      properties:
        accountType:
          type: string
          enum: [agent, host, developer, manager, other]
        country:
          type: string
        whatsappNumber:
          type: string
      required:
        - accountType
...
    NewsletterSubscribeBody:
      type: object
      properties:
        email:
          type: string
      required:
        - email